For #40549, Updated tk-config-basic branch versions (#6)
This adds command grouping functionality in tk-desktop and tk-multi-launch command. Rather than seeing one launch button per version of a DCC application, the end user should see one launch button for the DCC, and drop-down menu items to select specific versions of the DCC. The highest version is currently used as the default in tk-desktop.