Testing the feel of reducing code duplication via subclassing the plugins
This is a new branch meant to start a discussion.

The logic for publishing, verison creation, etc is now moved to base class
plugins. Implementations for the maya workflow have been done. We should
discuss this in terms of what it means for a TD looking at how the basic
plugins work and trying to customize the behavior. We need to balance that
with support and maintainability on our end. We also need to consider if it
makes sense to move some of the shared logic to an engine.

The subclassing works well but it also makes the hook paths in the
configuration quite long. Something else to look at together.

I also made one more change to add an "ensure session is saved" plugin that
runs first and is disabled (can't be unchecked). This prevents each of the
subsequent plugins from having to ensure the session is saved and further
reduces code duplication. This further reduces the publishing process to a
series of relatively atomic actions. We should also discuss this as a new
addition to the workflow.