Various bug fixes
Updated:
tk-core to v0.18.97
- Configurations now have transaction markers that indicate when an
update started and ended. If the ended marker is missing, the
configuration will be updated again since it is in an incoherent state.
There was also a performance issue with resolve_descriptor which wrote
the configuration to disk when it wasn't needed.

tk-desktop to v2.3.6
- Fixes an error when there is no pipeline configuration for a given
project.

tk-frameworkdesktopserver to v1.1.14
- Fixes an issue with the latest Toolkit core which fixes a bug where
classic configurations were not showing up in actions menu