Optimizations and bug fixes.
tk-core v0.18.114
- Path cache optimizations. Removed inner loop for secondary entities
and find queries are batched into smaller chunks. [44807]
- Previously, when an unhandled exception bubbles up, we replied to the
web app with the full exception stack. This fix ensures we reply with a
general error message and log the full exception. If debug logging is
turned on, we do reply with the full exception. [42770]
- unregister_folders command failed when unregistered a large amount of
folders. This fix resolves that issue by batching up any large deletions
into batched sql queries where 200 items are deleted at a time. [44931]

tk-framework-desktopserver
- Defaults to bootstrap for getting actions. The "legacy workaround" is
still in place, but is now behind an environment variable. [44081]

tk-multi-loader2
- Nuke read node frame range was not good when loaded from the Loader2
app. This fix ensures we correctly determine the start/end frame of an
image sequence. [44569]