tk-core update
Makes code parsable by the Python3 interpreter, fixes hook-related bug,
adds a bundle cache environment variable, adds a new metrics endpoint.

Details:

- Updates all tk-core code to be parsable with a Python 3 interpreter.
Note that this does not mean Toolkit is Python 3 compatible, it only
means that the code is lexically correct. This check however, is not
applied to the third-party packages yet.

- Fixes a bug related to template-type settings controlled by core
hooks. Previously, if one of these settings was configured to used a
hook to determine its value it would fail complaining about an undefined
variable.

- Adds support for the SHOTGUN_BUNDLE_CACHE_PATH environment variable
that allows one to specify the location of the path on disk (possibly
shared) to which all bundles would be downloaded to and made available
from.

- Revisits metrics logging with new endpoint.